Shorten name due to char limits on some objects
It seems like for some objects we are getting some char limit
restrictions. Even though having the component in the name of the
PipelineRun and in the config map might be useful for debugging
reasons, we cannot really be sure that it won't contain a too
long character. Due to this reason, let's remove the component
name.

Signed-off-by: Giulia Naponiello <gnaponie@redhat.com>