Remove timestamp field from component struct
The timestamp was previously used for implementing the refreshing of
GitHub App installation cache. Since the cache is now refreshed based
on time, we no longer need to store timestamp information in the
component struct.

Additionally, changed the timestamp in PipelineRun name from Unix
timestamp format to MMDDhhmm, making it more human-readable.