Introduce a new event reconciler to generate tokens on-demand
Previously, we generated Renovate tokens for components while creating
the pipelineruns, then before starting the pipelineruns for GitHub
components, we checked and refreshed the token to ensure it's valid for
the pipelinerun execution.

However, a queue system will be added to Konflux, meaning when we start
a pipelinerun, it may stay in the queue until the cluster is able to
execute it. We can't predict when the pipelinerun will actually start,
so refreshing the token before starting a pipelinerun can't ensure the
token is still valid when the pipelinerun is actually executed.

In this change, we don't create the renovate token in secret while
creating pipelineruns for GitHub components. Instead, we introduced a
new event reconciler to monitor events. When a pipelinerun is started
but the pod can't mount the renovate token (because we didn't create
it), an event will be triggered, with a message like this:

    MountVolume.SetUp failed for volume "secret-renovate-07110345-e8d4c7bf"
    : references non-existent secret key: renovate-token

The pod will retry and wait until the secret key `renovate-token` is
available. We will then generate the token and update the secret to make
the pod able to start. This ensures the pod has a valid GitHub token
when it's actually started.

JIRA: CWFHEALTH-4054