Remove pending status and concurrent limit for PipelineRuns
The upcoming queue system in Konflux cannot properly handle PipelineRuns
with a pending status. Therefore, we need to update the controller to
create PipelineRuns without setting them to pending status initially.

Previously, setting the pending status was used to control the maximum
number of concurrently running PipelineRuns. However, with the introduction
of the queue system, we will not be able to control that anymore.
Consequently, the related code (which moves PipelineRuns from pending
to running status) also needs to be removed.